chore: moved types to existing types folder
