+++ client/app/components/routes/operations/form/Operation.tsx
@@ -5,7 +5,7 @@ import { operationSchema } from "@/app/utils/jsonSchema/operations";
 import { RJSFSchema } from "@rjsf/utils";
 import { fetchAPI } from "@/app/utils/api";
 import Review from "./Review";
-import { Status } from "@/app/types";
+import { Status } from "@/app/types/types";
 
 // 🛠️ Function to fetch NAICS codes
 async function getNaicsCodes() {
@@ -39,19 +39,19 @@ async function getOperation(id: number) {
 export const createOperationSchema = (
   schema: RJSFSchema,
   naicsCodes: { id: number }[],
-  naicsCategories: { id: number }[],
+  naicsCategories: { id: number }[]
 ) => {
   const localSchema = JSON.parse(JSON.stringify(schema));
   // naics codes
   if (Array.isArray(naicsCodes)) {
     localSchema.properties.naics_code_id.enum = naicsCodes.map(
-      (code) => code.id,
+      (code) => code.id
     );
   }
   // naics categories
   if (Array.isArray(naicsCategories)) {
     localSchema.properties.naics_category_id.enum = naicsCategories.map(
-      (category) => category.id,
+      (category) => category.id
     );
   }
   return localSchema;
